Node 2
Clock og interrupt lagt til, Interrupt på PID funker ikke

+++ Node2_ny/Node2_ny/control_driver.h
@@ -14,13 +14,13 @@
 void CD_init();
 void CD_set_ref_pos(unsigned char pos);
 void CD_clk_init();
-double CD_read_encoder();
+int16_t CD_read_encoder();
 void CD_direc(unsigned char direc);
 void CD_speed(int value);
 void CD_velocity(int vel);
 unsigned char CD_pid_gain(float p,float i,float d);
 void CD_PID(int reference_value);
 void CD_encoder_reset();
-int CD_encoder_max();
+int16_t CD_encoder_max();
 
 #endif
